# apps/server/gen/openapi/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/auth/login:
        post:
            tags:
                - AuthService
            operationId: AuthService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/auth/register:
        post:
            tags:
                - AuthService
            operationId: AuthService_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RegisterResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/auth/server-info:
        get:
            tags:
                - AuthService
            operationId: AuthService_ServerInfo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ServerInfoResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        LoginRequest:
            type: object
            properties:
                userName:
                    type: string
                password:
                    type: string
        LoginResponse:
            type: object
            properties:
                token:
                    type: string
                userId:
                    type: string
                userName:
                    type: string
                role:
                    type: string
...
        RegisterRequest:
            type: object
            properties:
                userName:
                    type: string
                    description: Display name for the new user. Must be unique across the server.
                password:
                    type: string
                    description: Plaintext password. Server hashes with argon2id. Min 8 chars.
                inviteCode:
                    type: string
                    description: Required when registration is closed and the users table is non-empty.
                serverName:
                    type: string
                    description: |-
                        First-user only: optional server identity fields. Ignored on subsequent
                         Register calls — owners change them later via UpdateServerSettings.
                serverDescription:
                    type: string
        RegisterResponse:
            type: object
            properties:
                token:
                    type: string
                    description: 'JWT (HS256). Caller stores it and presents as Authorization: Bearer.'
                userId:
                    type: string
                userName:
                    type: string
                role:
                    type: string
...
        ServerInfoResponse:
            type: object
            properties:
                hasUsers:
                    type: boolean
                    description: |-
                        True once the server has at least one registered user. When false, the
                         next Register call promotes the caller to owner regardless of invite/
                         registration policy.
                registrationOpen:
                    type: boolean
                    description: |-
                        Mirrors the registration_open server setting. When true, Register works
                         without an invite code (only meaningful once has_users is true).
                version:
                    type: string
                    description: |-
                        Server build identity, surfaced unauthenticated so a CLI can warn on
                         CLI/server version skew before login.
                commit:
                    type: string
                    description: |-
                        Short commit sha for the running server build. May carry a "-dirty"
                         suffix for local builds.
                serverName:
                    type: string
                    description: Owner-set display name. Empty until configured.
                serverDescription:
                    type: string
                    description: Owner-set one-liner description. Empty until configured.
                ownerName:
                    type: string
                    description: Display name of the server owner. Empty until first Register completes.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...